pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0"]
build-backend = "setuptools.build_meta"

[project]
name = "dyna-automator"
version = "0.3.0"
authors = [
  { name="NoxelEcnord", email="derrickom005@gmail.com" },
]
description = "A modern, interactive web automation and scraping framework built with Playwright."
readme = "README.md"
requires-python = ">=3.8"
keywords = ["web-scraping", "automation", "playwright", "javascript", "cli", "interactive", "repl"]
dependencies = [
    "typer[all]>=0.9.0",
    "playwright>=1.44.0",
    "rich>=13.7.0",
    "prompt_toolkit>=3.0.0",
    "pyyaml>=6.0",
]

[project.urls]
"Homepage" = "https://github.com/NoxelEcnord/dyna-automator"
"Bug Tracker" = "https://github.com/NoxelEcnord/dyna-automator/issues"

[project.scripts]
dyna = "dyna_automator.cli.main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["dyna_automator*"]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]
